requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# AI/ML
ultralytics
torch
torchvision
opencv-python
opencv-contrib-python
numpy
pillow
scipy

# Database
sqlalchemy==2.0.25
psycopg2-binary
alembic==1.13.1

# Utilities
python-dotenv==1.0.1
pydantic
pydantic-settings

# Image Processing
scikit-image
matplotlib==3.8.2

# Additional
python-jose==3.3.0
passlib==1.7.4
python-dateutil==2.8.2